install: move IgnitionHash to io.rs
It's not all that specific to installation.
